requirements.txt
flask==3.0.0
crewai==0.11.0
python-dotenv==1.0.0
PyGithub==2.1.1
openai==1.3.7
